requirements.txt
click==7.0
ruamel.yaml==0.16.5
numpy==1.16.2
scipy==1.3.1
scikit-learn==0.21.2
gensim==3.4.0
torch==1.0.1
nltk==3.4
tqdm==4.31.1
joblib==0.13.2
logzero==1.5.0
